Factor out diagnostic slug checking from `DiagnosticDerive` and `LintDiagnosticDerive`
